+++ tests/apps/courses/test_models_page_role.py
@@ -26,15 +26,16 @@ class PageRoleModelsTestCase(TestCase):
 
         self.assertEqual(str(role), "Admin | My page")
 
-    def test_models_page_role_django_permissions_group_admin(self, *_):
+    def test_models_page_role_group_and_folder(self, *_):
         """
-        A group should be automatically created the first time the page role is saved.
+        A group and a folder should be automatically created the first time the page role is
+        saved.
         """
         page = PageFactory(title__title="My page")
         role = PageRole(page=page, role="ADMIN")
 
         with self.assertRaises(ObjectDoesNotExist):
-            self.assertEqual(role.folder._meta.model, Folder)
+            self.assertEqual(role.group._meta.model, Group)
 
         with self.assertRaises(ObjectDoesNotExist):
             self.assertEqual(role.folder._meta.model, Folder)
@@ -47,3 +48,18 @@ class PageRoleModelsTestCase(TestCase):
 
         self.assertEqual(role.group.name, "Admin | My page")
         self.assertEqual(role.folder.name, "Admin | My page")
+
+    def test_models_page_role_name_conflict(self, *_):
+        """
+        Creating page roles for 2 pages sharing the same title should not fail and should add
+        the id of the second page to the name of its group and folder to make it unique.
+        """
+        page1, page2 = PageFactory.create_batch(2, title__title="My page")
+        role1 = PageRole.objects.create(page=page1, role="ADMIN")
+        role2 = PageRole.objects.create(page=page2, role="ADMIN")
+
+        self.assertEqual(role1.group.name, "Admin | My page")
+        self.assertEqual(role1.folder.name, "Admin | My page")
+
+        self.assertEqual(role2.group.name, "Admin | My page [{:d}]".format(page2.id))
+        self.assertEqual(role2.folder.name, "Admin | My page [{:d}]".format(page2.id))